Dealing with UTC date formats, Notifying users when they receive a message, Subscribing to Route Parameter Changes, Optimizing the Presence, Optimizing the Messages

+++ demo/DatingApp/API/SignalR/MessageHub.cs
@@ -8,12 +8,12 @@ using Microsoft.AspNetCore.SignalR;
 namespace API.SignalR;
 
 public class MessageHub(IMessageRepository messageRepository, IUserRepository userRepository,
-    IMapper mapper) : Hub
+    IMapper mapper, IHubContext<PresenceHub> presenceHub) : Hub
 {
     public override async Task OnConnectedAsync()
     {
         var httpContext = Context.GetHttpContext();
-        var otherUser = httpContext.Request.Query["user"];
+        var otherUser = httpContext?.Request.Query["user"];
 
         if (Context.User == null || string.IsNullOrEmpty(otherUser))
             throw new Exception("Cannot join group");
@@ -21,17 +21,20 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
 
         await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
 
-        await AddToGroup(groupName);
+        var group = await AddToGroup(groupName);
+
+        await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
 
         var messages = await messageRepository.GetMessageThread(Context.User.GetUsername(), otherUser!);
 
-        await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+        await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
 
     }
 
     public override async Task OnDisconnectedAsync(Exception? exception)
     {
-        await RemoveFromMessageGroup();
+        var group = await RemoveFromMessageGroup();
+        await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
         await base.OnDisconnectedAsync(exception);
     }
 
@@ -65,6 +68,15 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
         {
             message.DateRead = DateTime.UtcNow;
         }
+        else
+        {
+            var connections = await PresenceTracker.GetConnectionsForUser(recipient.UserName);
+            if (connections != null && connections?.Count != null)
+            {
+                await presenceHub.Clients.Clients(connections).SendAsync("NewMessageReceived",
+                    new { username = sender.UserName, knownAs = sender.KnownAs });
+            }
+        }
 
 
         messageRepository.AddMessage(message);
@@ -77,7 +89,7 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
 
     }
 
-    private async Task<bool> AddToGroup(string groupName)
+    private async Task<Group> AddToGroup(string groupName)
     {
         var username = Context.User?.GetUsername() ?? throw new Exception("Cannot get username");
         var group = await messageRepository.GetMessageGroup(groupName);
@@ -91,17 +103,22 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
 
         group.Connections.Add(connection);
 
-        return await messageRepository.SaveAllAsync();
+        if (await messageRepository.SaveAllAsync()) return group;
+
+        throw new HubException("Failed to join group");
     }
 
-    private async Task RemoveFromMessageGroup()
+    private async Task<Group> RemoveFromMessageGroup()
     {
-        var connection = await messageRepository.GetConnection(Context.ConnectionId);
-        if (connection != null)
+        var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+        var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+        if (connection != null && group != null)
         {
             messageRepository.RemoveConnection(connection);
-            await messageRepository.SaveAllAsync();
+            if (await messageRepository.SaveAllAsync()) return group;
         }
+
+        throw new Exception("Failed to remove from group");
     }
 
     private string GetGroupName(string caller, string? other)
